refactor: separate unit and integration tests
