+++ agents/.coveragerc
@@ -1,6 +1,7 @@
 [run]
 omit =
     agents/dev/*
+    tests/integration/*
 
 [html]
 directory = agents/htmlcov
